requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.3
djangorestframework==3.15.2
gunicorn==21.2.0
packaging==26.0
psycopg2-binary==2.9.6
sqlparse==0.5.5
typing_extensions==4.13.2
django-filter
